update join to use hash_partition
