+++ ddf_library/functions/ml/feature_lib/extraction.py
@@ -30,13 +30,11 @@ class CountVectorizer(ModelDDF):
     >>> ddf2 = cv.transform(ddf1)
     """
 
-    def __init__(self, input_col, output_col=None, vocab_size=-1, min_tf=1.0,
-                 min_df=1.0, binary=True):
+    def __init__(self, input_col, vocab_size=200, min_tf=1.0,
+                 min_df=1, binary=True):
         """
         :param input_col: Input column name with the tokens;
-        :param output_col: Output field (default, add suffix '_vectorized');
-        :param vocab_size: Maximum size of the vocabulary. If -1, no limits
-         will be applied. (default, -1)
+        :param vocab_size: Maximum size of the vocabulary. (default, 200)
         :param min_tf: Specifies the minimum number of different documents a
          term must appear in to be included in the vocabulary. If this is an
          integer >= 1, this specifies the number of documents the term must
@@ -50,21 +48,17 @@ class CountVectorizer(ModelDDF):
         super(CountVectorizer, self).__init__()
 
         if isinstance(input_col, list):
-            raise Exception('`input_col` must be a single column')
-
-        if not output_col:
-            output_col = "{}_vectorized".format(input_col)
+            raise Exception('"input_col" must be a single column.')
 
         self.settings = dict()
-        self.settings['input_col'] = [input_col]
-        self.settings['output_col'] = output_col
+        self.settings['input_col'] = input_col
         self.settings['vocab_size'] = vocab_size
         self.settings['min_tf'] = min_tf
         self.settings['min_df'] = min_df
         self.settings['binary'] = binary
 
-        self.model = []
         self.name = 'CountVectorizer'
+        self.model = dict()
 
     def fit(self, data):
         """
@@ -85,37 +79,43 @@ class CountVectorizer(ModelDDF):
             result_p[f] = _wordcount(df[f], self.settings)
         word_dic = merge_reduce(merge_wordCount, result_p)
 
-        vocabulary = create_vocabulary(word_dic)
+        vocabulary = create_vocabulary(word_dic, -1)
 
         if any([min_tf > 0, min_df > 0, vocab_size > 0]):
             vocabulary = filter_words(vocabulary, self.settings)
 
-        self.model = [vocabulary]
+        self.model['algorithm'] = self.name
+        self.model['vocabulary'] = vocabulary
 
         return self
 
-    def fit_transform(self, data):
+    def fit_transform(self, data, output_col=None):
         """
         Fit the model and transform.
 
         :param data: DDF
+        :param output_col: Output field (default, add suffix '_vectorized');
         :return: DDF
         """
 
         self.fit(data)
-        ddf = self.transform(data)
+        ddf = self.transform(data, output_col)
 
         return ddf
 
-    def transform(self, data):
+    def transform(self, data, output_col=None):
         """
         :param data: DDF
+        :param output_col: Output field (default, add suffix '_vectorized');
         :return: DDF
         """
         if len(self.model) == 0:
             raise Exception("Model is not fitted.")
 
-        vocabulary = self.model[0]
+        if output_col is not None:
+            self.settings['output_col'] = output_col
+
+        vocabulary = self.model['vocabulary']
 
         df, nfrag, tmp = self._ddf_inital_setup(data)
 
@@ -139,12 +139,12 @@ class CountVectorizer(ModelDDF):
 def _wordcount(data, params):
     """Auxilar method to create a model."""
     wordcount = {}
-    column = params['input_col']
+    columns = params['input_col']
     # first:   Number of all occorrences with term t
     # second:  Number of diferent documents with term t
     # third:   temporary - only to idetify the last occorrence
 
-    for i_doc, doc in enumerate(data[column].values):
+    for i_doc, doc in enumerate(data[columns].tolist()):
 
         for token in doc:
             if token not in wordcount:
@@ -177,13 +177,27 @@ def merge_lists(list1, list2):
 
 
 # @local
-def create_vocabulary(word_dic):
+def create_vocabulary(word_dic, n_rows=-1):
     """Create a partial mode."""
     word_dic = compss_wait_on(word_dic)
-    docs_list = [[i[0], i[1][0], i[1][1]] for i in word_dic.items()]
+
     names = ['Word', 'TotalFrequency', 'DistinctFrequency']
-    voc = pd.DataFrame(docs_list, columns=names)\
-        .sort_values(by=['Word'])
+    docs_list = [[i[0], i[1][0], i[1][1]] for i in word_dic.items()]
+    voc = pd.DataFrame(docs_list, columns=names).sort_values(by=['Word'])
+
+    if n_rows != -1:
+        n_rows = compss_wait_on(n_rows)
+
+        # smooth_idf: Smooth idf weights by adding one to document
+        # frequencies, as if an extra document was seen containing every term in
+        # the collection exactly once.Prevents zero divisions.
+
+        # IDF = log_e(Total number of documents /
+        #            Number of documents containing t).
+        voc['InverseDistinctFrequency'] = \
+            np.log((n_rows + 1)/(voc['DistinctFrequency']+1)) + 1
+
+    voc.reset_index(drop=True, inplace=True)
     return voc
 
 
@@ -205,34 +219,49 @@ def filter_words(vocabulary, params):
 
 @task(returns=2)
 def _transform_bow(data, vocabulary, params, frag):
-    output_col = params['output_col']
-    column = params['input_col']
-    binary = params['binary']
 
-    vocabulary = vocabulary['Word'].values
+    vocabulary_ = {k: v for v, k in enumerate(vocabulary['Word'].values)}
+    n_cols = len(vocabulary_)
+    column = params['input_col']
+    output_col = params.get('output_col',
+                            ['vec_{}'.format(i) for i in range(n_cols)])
+
+    if not isinstance(output_col, list):
+        output_col = ['{}{}'.format(output_col, i) for i in range(n_cols)]
+
     data.reset_index(drop=True, inplace=True)
 
-    if len(data) > 0:
-        vector = np.zeros((len(data), len(vocabulary)), dtype=np.int)
-        for i, lines in data[column].values:
-            for e, w in enumerate(vocabulary):
-                if binary:
-                    if w in lines:
-                        vector[i][e] = 1
-                    else:
-                        vector[i][e] = 0
-                else:
-                    vector[i][e] = (lines == w).sum()
+    def dummy_fun(doc):
+        return doc
 
-        vector = vector.tolist()
+    from sklearn.feature_extraction.text import CountVectorizer
+    vectorizer = CountVectorizer(analyzer='word',
+                                 tokenizer=dummy_fun,
+                                 preprocessor=dummy_fun,
+                                 token_pattern=None,
+                                 binary=params['binary'])
+    vectorizer.vocabulary_ = vocabulary_
+    vector = vectorizer.transform(data[column].tolist())
 
-    else:
-        vector = np.nan
+    # vector = np.zeros((len(data), len(vocabulary)), dtype=np.int)
+    # if len(data) > 0:
+    #     for i, lines in enumerate(data[column].tolist()):
+    #         for e, w in enumerate(vocabulary):
+    #             if binary:
+    #                 if w in lines:
+    #                     vector[i][e] = 1
+    #                 else:
+    #                     vector[i][e] = 0
+    #             else:
+    #                 vector[i][e] = (lines == w).sum()
 
-    if output_col in data.columns:
-        ddata.drop([output_col], axis=1, inplace=True)
+    cols = [col for col in output_col if col in data.columns]
+    if len(cols) > 0:
+        data.drop(cols, axis=1, inplace=True)
 
-    df[output_col] = vector
+    vector = pd.DataFrame(vector.toarray(), columns=output_col)
+
+    data = pd.concat([data, vector], sort=False, axis=1)
 
     info = generate_info(data, frag)
     return data, info
@@ -250,13 +279,10 @@ class TfidfVectorizer(ModelDDF):
     >>> ddf2 = tfidf.transform(ddf1)
     """
 
-    def __init__(self, input_col, output_col=None, vocab_size=-1, min_tf=1.0,
-                 min_df=1.0):
+    def __init__(self, input_col, vocab_size=200, min_tf=1.0, min_df=1):
         """
         :param input_col: Input column name with the tokens;
-        :param output_col: Output field (default, add suffix '_vectorized');
-        :param vocab_size: Maximum size of the vocabulary. If -1, no limits
-         will be applied. (default, -1)
+        :param vocab_size: Maximum size of the vocabulary. (default, 200)
         :param min_tf: Specifies the minimum number of different documents a
          term must appear in to be included in the vocabulary. If this is an
          integer >= 1, this specifies the number of documents the term must
@@ -269,19 +295,15 @@ class TfidfVectorizer(ModelDDF):
         super(TfidfVectorizer, self).__init__()
 
         if isinstance(input_col, list):
-            raise Exception('`input_col` must be a single column')
-
-        if not output_col:
-            output_col = "{}_vectorized".format(input_col)
+            raise Exception('"input_col" must be a single column.')
 
         self.settings = dict()
-        self.settings['input_col'] = [input_col]
-        self.settings['output_col'] = output_col
+        self.settings['input_col'] = input_col
         self.settings['vocab_size'] = vocab_size
         self.settings['min_tf'] = min_tf
         self.settings['min_df'] = min_df
 
-        self.model = []
+        self.model = dict()
         self.name = 'TfidfVectorizer'
 
     def fit(self, data):
@@ -298,52 +320,60 @@ class TfidfVectorizer(ModelDDF):
 
         df, nfrag, tmp = self._ddf_inital_setup(data)
 
+        # TODO: info instead to generate new tasks
+        counts = [count_records(df[f]) for f in range(nfrag)]
+        count = merge_reduce(mergeCount, counts)
+
         result_p = [[] for _ in range(nfrag)]
         for f in range(nfrag):
             result_p[f] = _wordcount(df[f], self.settings)
         word_dic = merge_reduce(merge_wordCount, result_p)
-        vocabulary = create_vocabulary(word_dic)
+        vocabulary = create_vocabulary(word_dic, n_rows=count)
 
         if any([min_tf > 0, min_df > 0, vocab_size > 0]):
             vocabulary = filter_words(vocabulary, self.settings)
 
-        self.model = [vocabulary]
+        self.model['vocabulary'] = vocabulary
+        self.model['algorithm'] = self.name
 
         return self
 
-    def fit_transform(self, data):
+    def fit_transform(self, data, output_col=None):
         """
         Fit the model and transform.
 
         :param data: DDF
+        :param output_col: Output field (default, add suffix '_vectorized');
         :return: DDF
         """
 
         self.fit(data)
-        ddf = self.transform(data)
+        ddf = self.transform(data, output_col)
 
         return ddf
 
-    def transform(self, data):
+    def transform(self, data, output_col=None):
         """
         :param data: DDF
+        :param output_col: Output field (default, add suffix '_vectorized');
         :return: DDF
         """
 
         if len(self.model) == 0:
             raise Exception("Model is not fitted.")
-        vocabulary = self.model[0]
+
+        if output_col is not None:
+            self.settings['output_col'] = output_col
+
+        vocabulary = self.model['vocabulary']
 
         df, nfrag, tmp = self._ddf_inital_setup(data)
-
-        counts = [count_records(df[f]) for f in range(nfrag)]
-        count = merge_reduce(mergeCount, counts)
 
         result = [[] for _ in range(nfrag)]
         info = [[] for _ in range(nfrag)]
         for f in range(nfrag):
             result[f], info[f] = \
-                construct_tf_idf(df[f], vocabulary, self.settings, count, f)
+                construct_tf_idf(df[f], vocabulary, self.settings, f)
 
         uuid_key = self._ddf_add_task(task_name='task_transform_tfidf',
                                       status='COMPLETED', lazy=False,
@@ -355,20 +385,20 @@ class TfidfVectorizer(ModelDDF):
         return DDF(task_list=tmp.task_list, last_uuid=uuid_key)
 
 
-@task(returns=list)
+@task(returns=1)
 def count_records(data):
     """Count the partial number of records in each fragment."""
     return len(data)
 
 
-@task(returns=list)
+@task(returns=1)
 def mergeCount(data1, data2):
     """Auxiliar method to merge the lengths."""
     return data1 + data2
 
 
 @task(returns=2)
-def construct_tf_idf(data, vocabulary, params, num_doc, frag):
+def construct_tf_idf(data, vocabulary, params, frag):
     """Construct the tf-idf feature.
 
     TF(t)  = (Number of times term t appears in a document)
@@ -378,36 +408,52 @@ def construct_tf_idf(data, vocabulary, params, num_doc, frag):
     Source: http://www.tfidf.com/
     """
 
-    alias = params['output_col']
-    columns = params['input_col']
-    vector = np.zeros((data.shape[0], vocabulary.shape[0]), dtype=np.float)
-    vocab = vocabulary['Word'].values
+    vocabulary_ = {k: v for v, k in enumerate(vocabulary['Word'].values)}
+    n_cols = len(vocabulary_)
+    column = params['input_col']
+    output_col = params.get('output_col',
+                            ['vec_{}'.format(i) for i in range(n_cols)])
+
+    if not isinstance(output_col, list):
+        output_col = ['{}{}'.format(output_col, i) for i in range(n_cols)]
+
     data.reset_index(drop=True, inplace=True)
 
-    for i, point in data.iterrows():
-        lines = point[columns].values
-        lines = np.array(list(itertools.chain(lines))).flatten()
-        for w, token in enumerate(vocab):
-            if token in lines:
-                # TF = (Number of times term t appears in the document) /
-                #        (Total number of terms in the document).
-                nTimesTermT = np.count_nonzero(lines == token)
-                total = len(lines)
-                if total > 0:
-                    tf = float(nTimesTermT) / total
-                else:
-                    tf = 0
+    def dummy_fun(doc):
+        return doc
 
-                # IDF = log_e(Total number of documents /
-                #            Number of documents with term t in it).
-                nDocsWithTermT = vocabulary.\
-                    loc[vocabulary['Word'] == token, 'DistinctFrequency'].\
-                    item()
-                idf = np.log(float(num_doc) / nDocsWithTermT)
+    from sklearn.feature_extraction.text import TfidfVectorizer
+    vectorizer = TfidfVectorizer(analyzer='word',
+                                 tokenizer=dummy_fun,
+                                 preprocessor=dummy_fun,
+                                 token_pattern=None)
+    vectorizer.vocabulary_ = vocabulary_
+    vectorizer.idf_ = vocabulary['InverseDistinctFrequency'].values
+    vector = vectorizer.transform(data[column].tolist())
 
-                vector[i][w] = tf*idf
+    # vector = np.zeros((data.shape[0], vocabulary.shape[0]), dtype=np.float)
+    # if len(data) > 0:
+    #     vocab = vocabulary['Word'].values
+    #     for i, lines in enumerate(data[column].tolist()):
+    #         total = len(lines)
+    #         for w, token in enumerate(vocab):
+    #             if token in lines:
+    #                 # TF = (Number of times term t appears in the document) /
+    #                 #        (Total number of terms in the document).
+    #                 tf = np.count_nonzero(lines == token) / total
+    #
+    #                 idf = vocabulary.\
+    #                     loc[vocabulary['Word'] == token,
+    #                         'InverseDistinctFrequency'].item()
+    #
+    #                 vector[i][w] = tf*idf
 
-    data[alias] = vector.tolist()
+    cols = [col for col in output_col if col in data.columns]
+    if len(cols) > 0:
+        data.drop(cols, axis=1, inplace=True)
+
+    vector = pd.DataFrame(vector.toarray(), columns=output_col)
+    data = pd.concat([data, vector], sort=False, axis=1)
 
     info = generate_info(data, frag)
     return data, info
